Textures added to folders Added truck, RND amd farmer sounds (13) dog added added Resources (font and UISkin) detaching actors baseActor Added some fixxs and button added All Screens and navegation added funcionality to sound and user( both remain in preferences) added Mock Added dialogs added spinActor and fixxed labels Funcionality on shop and store Added backgrounds and animal functionality

+++ core/src/com/mygdx/game/Dominio/Precio.java
@@ -4,9 +4,9 @@ public class Precio {
 
     private int id;
     private TipoRecurso tipoRecurso;
-    private float precio;
+    private int precio;
 
-    public Precio(int id, TipoRecurso tipoRecurso, float precio) {
+    public Precio(int id, TipoRecurso tipoRecurso, int precio) {
         this.id = id;
         this.tipoRecurso = tipoRecurso;
         this.precio = precio;
@@ -28,11 +28,11 @@ public class Precio {
         this.tipoRecurso = tipoRecurso;
     }
 
-    public float getPrecio() {
+    public int getPrecio() {
         return precio;
     }
 
-    public void setPrecio(float precio) {
+    public void setPrecio(int precio) {
         this.precio = precio;
     }
 }
